+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
  * Added SPBlobCacheSettings resource
  * Fixed a bug with SPTimerJobState that prevented a custom schedule being applied to a timer job
+ * Added SPOfficeOnlineServerBinding resource
+ * Fixed a bug with the detection of group principals vs. user principals in SPServiceAppSecurity
 
 ### 1.0